pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "integritymark"
version = "0.1.0"
description = "Research code for speech watermark training, evaluation, detection, and streaming deployment."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "IntegrityMark Project Team" }
]
dependencies = [
  "audiomentations>=0.34",
  "einops>=0.7",
  "hydra-core>=1.3",
  "julius>=0.2.7",
  "librosa>=0.9",
  "matplotlib>=3.7",
  "numpy>=1.23,<2",
  "omegaconf>=2.3",
  "pandas>=2.0",
  "pesq>=0.0.4",
  "pystoi>=0.4.1",
  "scipy>=1.10",
  "sounddevice>=0.4",
  "torch>=2.1",
  "torchaudio>=2.1",
  "tqdm>=4.66",
  "wandb>=0.16",
  "websockets>=12",
]

[project.optional-dependencies]
stream = [
  "sounddevice>=0.4",
  "websockets>=12",
]
dev = [
  "ipykernel>=6.29",
]

[project.scripts]
integritymark-detect = "detrect:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["dataset*", "distortions*", "losses*", "models*", "streaming*", "utils*"]
